rename the directory 'pages' to 'views'
