+++ src/views/registration-page/index.js
